feature(reactions): Unify reaction messages with emoji status and pass it through akka
