+++ bigbluebutton-html5/imports/api/user-reaction/server/eventHandlers.js
@@ -1,0 +1,4 @@
+import RedisPubSub from '/imports/startup/server/redis';
+import handleSetUserReaction from './handlers/setUserReaction';
+
+RedisPubSub.on('UserEmojiChangedEvtMsg', handleSetUserReaction);
