+++ bigbluebutton-html5/imports/api/user-reaction/server/index.js
@@ -1,2 +1,3 @@
+import './eventHandlers';
 import './methods';
 import './publishers';
