+++ bigbluebutton-html5/imports/api/user-reaction/server/methods/setUserReaction.js
@@ -1,13 +1,32 @@
+import { Meteor } from 'meteor/meteor';
 import { check } from 'meteor/check';
+import RedisPubSub from '/imports/startup/server/redis';
+import Logger from '/imports/startup/server/logger';
 import { extractCredentials } from '/imports/api/common/server/helpers';
-import addUserReaction from '/imports/api/user-reaction/server/modifiers/addUserReaction';
 
-export default function setUserReaction(reaction) {
-  check(reaction, String);
-  const { meetingId, requesterUserId } = extractCredentials(this.userId);
+export default function setUserReaction(emoji) {
+  try {
+    const REDIS_CONFIG = Meteor.settings.private.redis;
+    const CHANNEL = REDIS_CONFIG.channels.toAkkaApps;
+    const EVENT_NAME = 'ChangeUserEmojiCmdMsg';
 
-  check(meetingId, String);
-  check(requesterUserId, String);
+    const { meetingId, requesterUserId } = extractCredentials(this.userId);
 
-  addUserReaction(meetingId, requesterUserId, reaction);
+    check(meetingId, String);
+    check(requesterUserId, String);
+    check(emoji, String);
+
+    const payload = {
+      emoji,
+      userId: requesterUserId,
+    };
+
+    Logger.verbose('User emoji status updated', {
+      emoji, requesterUserId, meetingId,
+    });
+
+    RedisPubSub.publishUserMessage(CHANNEL, EVENT_NAME, meetingId, requesterUserId, payload);
+  } catch (err) {
+    Logger.error(`Exception while invoking method setUserReaction ${err.stack}`);
+  }
 }
