+++ bigbluebutton-html5/imports/api/users/server/eventHandlers.js
@@ -4,7 +4,6 @@ import handleUserJoined from './handlers/userJoined';
 import handleUserLeftFlagUpdated from './handlers/userLeftFlagUpdated';
 import handleValidateAuthToken from './handlers/validateAuthToken';
 import handlePresenterAssigned from './handlers/presenterAssigned';
-import handleEmojiStatus from './handlers/emojiStatus';
 import handleChangeRole from './handlers/changeRole';
 import handleUserPinChanged from './handlers/userPinChanged';
 import handleUserInactivityInspect from './handlers/userInactivityInspect';
@@ -14,7 +13,6 @@ RedisPubSub.on('PresenterAssignedEvtMsg', handlePresenterAssigned);
 RedisPubSub.on('UserJoinedMeetingEvtMsg', handleUserJoined);
 RedisPubSub.on('UserLeftMeetingEvtMsg', handleRemoveUser);
 RedisPubSub.on('ValidateAuthTokenRespMsg', handleValidateAuthToken);
-RedisPubSub.on('UserEmojiChangedEvtMsg', handleEmojiStatus);
 RedisPubSub.on('UserRoleChangedEvtMsg', handleChangeRole);
 RedisPubSub.on('UserMobileFlagChangedEvtMsg', handleChangeMobileFlag);
 RedisPubSub.on('UserLeftFlagUpdatedEvtMsg', handleUserLeftFlagUpdated);
