+++ bigbluebutton-html5/imports/api/users/server/methods.js
@@ -1,6 +1,5 @@
 import { Meteor } from 'meteor/meteor';
 import validateAuthToken from './methods/validateAuthToken';
-import setEmojiStatus from './methods/setEmojiStatus';
 import setSpeechLocale from './methods/setSpeechLocale';
 import setMobileUser from './methods/setMobileUser';
 import assignPresenter from './methods/assignPresenter';
@@ -15,7 +14,6 @@ import setRandomUser from './methods/setRandomUser';
 import setExitReason from './methods/setExitReason';
 
 Meteor.methods({
-  setEmojiStatus,
   setSpeechLocale,
   setMobileUser,
   assignPresenter,
