--- bigbluebutton-html5/imports/api/users/server/methods/setEmojiStatus.js
@@ -1,33 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-import { check } from 'meteor/check';
-import RedisPubSub from '/imports/startup/server/redis';
-import Logger from '/imports/startup/server/logger';
-import { extractCredentials } from '/imports/api/common/server/helpers';
-
-export default function setEmojiStatus(userId, status) {
-  try {
-    const REDIS_CONFIG = Meteor.settings.private.redis;
-    const CHANNEL = REDIS_CONFIG.channels.toAkkaApps;
-    const EVENT_NAME = 'ChangeUserEmojiCmdMsg';
-
-    const { meetingId, requesterUserId } = extractCredentials(this.userId);
-
-    check(meetingId, String);
-    check(requesterUserId, String);
-    check(userId, String);
-    check(status, String);
-
-    const payload = {
-      emoji: status,
-      userId,
-    };
-
-    Logger.verbose('User emoji status updated', {
-      userId, status, requesterUserId, meetingId,
-    });
-
-    RedisPubSub.publishUserMessage(CHANNEL, EVENT_NAME, meetingId, requesterUserId, payload);
-  } catch (err) {
-    Logger.error(`Exception while invoking method setEmojiStatus ${err.stack}`);
-  }
-}
